resource-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0

sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
psycopg2-binary==2.9.10

pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

google-genai
neo4j==6.1.0
